+++ src/mathematics/uncertainvalues/add_uncertainvalues.jl
@@ -129,20 +129,20 @@ end
 # Special cases 
 #####################################################################################
 
-import ..UncertainValues: CertainValue
+import ..UncertainValues: CertainScalar
 
 ##################
-# `CertainValue`s
+# `CertainScalar`s
 #################
 """
-    Base.:+(a::Union{CertainValue, Real}, b::Union{CertainValue, Real})
+    Base.:+(a::Union{CertainScalar, Real}, b::Union{CertainScalar, Real})
 
 Addition of certain values with themselves or scalars acts as regular addition, but 
-returns the result wrapped in a `CertainValue` instance.
+returns the result wrapped in a `CertainScalar` instance.
 """
-Base.:+(a::Union{CertainValue, Real}, b::Union{CertainValue, Real}) 
+Base.:+(a::Union{CertainScalar, Real}, b::Union{CertainScalar, Real}) 
 
-Base.:+(a::CertainValue, b::CertainValue) = CertainValue(a.value + b.value)
-Base.:+(a::CertainValue, b::Real) = CertainValue(a.value + b)
-Base.:+(a::Real, b::CertainValue) = CertainValue(a + b.value)
+Base.:+(a::CertainScalar, b::CertainScalar) = CertainScalar(a.value + b.value)
+Base.:+(a::CertainScalar, b::Real) = CertainScalar(a.value + b)
+Base.:+(a::Real, b::CertainScalar) = CertainScalar(a + b.value)