+++ src/uncertain_values/convert.jl
@@ -1,2 +1,2 @@
-convert(::Type{CertainValue}, x::T) where {T <: Number} = CertainValue(x)
-convert(::Type{T1}, x::T2) where {T1 <: AbstractUncertainValue, T2 <: Number} = CertainValue(x)
+convert(::Type{CertainScalar}, x::T) where {T <: Number} = CertainScalar(x)
+convert(::Type{T1}, x::T2) where {T1 <: AbstractUncertainValue, T2 <: Number} = CertainScalar(x)