+++ test/resampling/uncertain_values/test_resampling_certain_value.jl
@@ -1,4 +1,4 @@
-x = CertainValue(2.0)
+x = CertainScalar(2.0)
 
 test_constraints = [
     NoConstraint(),