+++ test/uncertain_values/populations/test_UncertainScalarPopulation.jl
@@ -1,6 +1,6 @@
 import StatsBase: AbstractWeights
 
-# Uncertain population consisting of CertainValues (scalars get promoted to CertainValue)s
+# Uncertain population consisting of CertainScalars (scalars get promoted to CertainScalar)s
 # theoretical distributions and KDE distributions
 p1 = UncertainScalarPopulation(
         [3.0, UncertainValue(Normal, 0, 1),