+++ test/uncertain_values/test_certain_values.jl
@@ -2,5 +2,5 @@
 x = 3
 y = 3.3 
 
-@test UncertainValue(x) isa CertainValue
-@test UncertainValue(y) isa CertainValue
+@test UncertainValue(x) isa CertainScalar
+@test UncertainValue(y) isa CertainScalar